+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## v1.0.0
+### Breaking Changes
+- The library returns now a result and will no longer panic
+  under certain circumstances.
+- removed `per_element_scaling_fn` in favor of
+  `complex_scaling_fn`, which is now just called
+  `scaling_fn` - the old behaviour was more confusing than
+  beneficial
+- renamed `ComplexSpectrumScalingFunction` to `SpectrumScalingFunction` and
+  moved it into the `scaling`-module
+- MSRV is now `1.52.1` stable
+### Other
+- many internal improvements
+- rust-toolchain.toml for build stability and reproducibility
+- really minor performance improvements (~5 %)
+
 ## v0.5.1
 - Feature "rustfft-complex" uses "rustfft"-crate at version 6 which is faster/more optimized (~25%).
 - improved CI